项目重构： 1.数据库效率和可靠性增加（外键和索引） 2.增加kafka集群，实现异步操作 3.重构网关实现直接转发 4.重构service方法 5.订单字段更加丰富 6.实现多语句一次执行

+++ consumer-feign-client4001/src/main/java/com/zxy/work/controller/PaymentServiceClientController.java
@@ -41,7 +41,7 @@ public class PaymentServiceClientController {
      * @return 支付删除结果
      */
     @DeleteMapping("/delete")
-    public ApiResponse<String> delete(@RequestParam("orderId") Integer orderId) throws MyException {
+    public ApiResponse<String> delete(@RequestParam("orderId") long orderId) throws MyException {
         log.info("删除支付：" + orderId);
         try{
             return paymentServiceClient.delete(orderId);
@@ -73,7 +73,7 @@ public class PaymentServiceClientController {
      * @return 查询结果
      */
     @GetMapping("/getByOrderId")
-    public ApiResponse<Object> getByOrderId(@RequestParam("orderId") Integer orderId) throws MyException {
+    public ApiResponse<Object> getByOrderId(@RequestParam("orderId") long orderId) throws MyException {
         log.info("通过订单id获取支付：" + orderId);
         try{
             return paymentServiceClient.getByOrderId(orderId);
