+++ provider-payment-service6001/src/main/java/com/zxy/work/dao/PaymentMapper.java
@@ -9,10 +9,10 @@ public interface PaymentMapper {
 
     int create(Payment payment);
 
-    int delete(@Param("orderId")Integer orderId);
+    int delete(@Param("orderId")long orderId);
 
     int update(Payment payment);
 
-    Payment selectByOrderId(@Param("orderId") Integer orderId);
+    Payment selectByOrderId(@Param("orderId")long orderId);
 
 }
